Type file refactor
