--- app/3d/types.ts
@@ -1,4 +1,0 @@
-import type { JSX } from "react";
-
-export type MeshProps = Readonly<JSX.IntrinsicElements["mesh"]>;
-export type GroupProps = Readonly<JSX.IntrinsicElements["group"]>;
